Stop on number of acceleration points rather than digitized points.
The number of digitized points reported in the component headers doesn't
always match the number of data points in the file. Instead of using
this number as the test condition for stopping reading data, use the
number of acceleration points which seems to be correct.

This is part of a fix for GitHub issue #2.