When parsing, break the lines into data points in a more robust manner.
Previously we were just using the .split() method on each line to break
it into the individual values. This works when the values are small
enough for there to be a space between them. However, GeoNet seems to
use 999999.9 as its NaN value which means there is no space between two
values.

The new way, chosen from several possible methods via some basic
timeit.timeit() profiling, is to prepare some slice indices ahead of
time, and use these to split the line up. We also remove empty values
(e.g., if the final line of data is not full).

GitHub issues: closes #2.