Add ability to realign values to epicentre bearing.
Rework the Record initialiser to take a parameter defining how to
realign the measurements. This parameter is taken from a sub-class
holding constants. It defaults to the north-and-east alignment of
previous versions. The new options are no realignment (i.e., use the
data straight from the file), or to realign to axes defined by the
bearing from the site to the epicentre and a 90 degrees clockwise
cross-bearing.

GitHub issues: closes #1.